Facebook auth

+++ aws/functions/deleteBrowseData/index.js
@@ -7,10 +7,9 @@
  * @resource: /browses
  * @method: DELETE
  * @params:
- *      - browser: username [string]
  *      - published: browse published timestamp [integer]
  *      - shot: shot URL [string]
- *      - token: jwt token from cognito [string]
+ *      - token: access token from facebook [string]
  * @returns:
  *      - browser [string]
  *      - url [string]
@@ -24,10 +23,6 @@ const request = require('request');
 
 
 exports.handle = function handler(event, context) {
-  if (!event.browser) {
-    context.fail('Bad Request: Missing browser parameter');
-    return;
-  }
   if (!event.published) {
     context.fail('Bad Request: Missing published parameter');
     return;
@@ -44,52 +39,55 @@ exports.handle = function handler(event, context) {
    * Validate JSON Web Token.
    */
   request({
-    url: 'https://7ibd5w7y69.execute-api.eu-west-1.amazonaws.com/beta/validate',
-    method: 'POST',
-    headers: {
-      'content-type': 'application/json',
-    },
-    json: { username: event.browser, token: event.token, service: 'browses' },
+    url: `https://graph.facebook.com/me?access_token=${event.token}`,
+    method: 'GET',
   }, (error, rsp, body) => {
     if (!error && rsp.statusCode === 200) {
-      // Successfully validated token
-      const browseParams = {
-        TableName: 'browses',
-        Key: {
-          browser: event.browser,
-          published: event.published,
-        },
-      };
-      if (event.shot.indexOf('browses/') <= -1) {
-        context.fail('Unprocessable Entity: Unrecognised shot url');
-        return;
-      }
-      const s3Params = {
-        Bucket: 'browses',
-        Key: event.shot.split('browses/')[1],
-      };
-      /*
-       * Delete browse data from browses table.
-       */
-      dynamo.delete(browseParams, (browseErr) => {
-        if (browseErr) {
-          context.fail('Internal Error: Failed to delete browse.');
+      if (!body.hasOwnProperty('error')) {
+        // Successfully validated token
+        const browser = JSON.parse(body).id;
+        const browseParams = {
+          TableName: 'browses',
+          Key: {
+            browser,
+            published: event.published,
+          },
+        };
+        if (event.shot.indexOf('browses/') <= -1) {
+          context.fail('Unprocessable Entity: Unrecognised shot url');
           return;
         }
-        s3.deleteObject(s3Params, (s3Err) => {
-          if (s3Err) {
-            context.fail('Internal Error: Failed to delete screenshot');
+        const s3Params = {
+          Bucket: 'browses',
+          Key: event.shot.split('browses/')[1],
+        };
+        /*
+         * Delete browse data from browses table.
+         */
+        dynamo.delete(browseParams, (browseErr) => {
+          if (browseErr) {
+            context.fail('Internal Error: Failed to delete browse.');
             return;
           }
-          context.succeed({
-            browser: event.browser,
-            shot: event.shot,
-            published: event.published,
+          s3.deleteObject(s3Params, (s3Err) => {
+            if (s3Err) {
+              context.fail('Internal Error: Failed to delete screenshot');
+              return;
+            }
+            context.succeed({
+              browser,
+              shot: event.shot,
+              published: event.published,
+            });
           });
         });
-      });
+      } else {
+        const resp = JSON.parse(body);
+        context.fail(`Unprocessable Entity: ${resp.error.message}`);
+        return;
+      }
     } else {
-      context.fail(body.errorMessage);
+      context.fail('Internal Error: Failed to authorise with Facebook');
       return;
     }
   });
